Added a function that gets the average score of a family in a given round.
We can't just use the round function, since that just gives us a sample size of 1. As a side note, I played around with this function a bit and found that a sample size of 5000 was reasonable when dealing with 8 questions and odds of 10%. Anybody adapting this function might want to keep that in mind.

--Matteo Raso